+++ minitrace-macro/tests/ui/err/has-name-and-short-name.rs
@@ -1,6 +1,6 @@
 use minitrace::trace;
 
-#[trace(name = "Name", path_name = true)]
+#[trace(name = "Name", short_name = true)]
 fn f() {}
 
 fn main() {}